Implementing the famous byte at a time attack on ECB
In the challenge we have control only over the input.Thus we manipulate the input and find out the sudden increase in size of the ciphertext.Now we have initial blocks containing our hidden text and input character and the final block containing the padding.We know the number of inputs given thus we can find out the length of the secret text by subtracting length of input from the total length(input+secret). Now we have the last character of the secret text at the end of the block(excluding the padding block). Now we just check for all possible characters and find the character whose encryption matches that of our last character.Then on increasing the input size,the second last character is pushed to the end and the before mentioned checking is carried out.This is continued till we get the initial character of the secret text.Now combining all the characters we get the hidden texts.